Use versioned extension cache entry, if we have...
a version. Otherwise the extension JNLP will be downloaded a 2nd time and
a second entry will be created:
::i=0\13::l=http://localhost/myapp/jnlp.extension/common.jnlp::a=1706698483984::
::i=0\0::l=http://localhost/myapp/jnlp.extension/common.jnlp::v=6.2.3::a=1706698483198::

Because the cache lookup is called twice on extensions. The second time without the version
so the versioned entry is not used. Instead a new entry i=0\13 without a version is created.

This works fine as long as not a new version of common.jnlp::v=6.2.4 is lookuped.
Then the sorting in line 246 of net/adoptopenjdk/icedteaweb/resources/cache/CacheImpl.java
fails withe a NPE, because it's missing the version.